[수정] 조회수 증가 api 추가에 따른 수정사항들 수정

+++ src/main/java/com/juno/appling/domain/product/entity/Product.java
@@ -40,7 +40,7 @@ public class Product {
     private String image1;
     private String image2;
     private String image3;
-    private int viewCnt;
+    private Long viewCnt;
     private LocalDateTime createAt;
     private LocalDateTime modifiedAt;
 
@@ -60,7 +60,7 @@ public class Product {
         this.image1 = image1;
         this.image2 = image2;
         this.image3 = image3;
-        this.viewCnt = 0;
+        this.viewCnt = 0L;
         this.createAt = createAt;
         this.modifiedAt = modifiedAt;
     }
